requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==2.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-debug-toolbar==2.1
django-js-asset==1.2.2
pytz==2021.1
soupsieve==1.9.5
sqlparse==0.4.1
